ugc-service/fastapi-ugc/pyproject.toml
[tool.poetry]
name = "UGC-service"
version = "0.1.0"
description = "First UGC-service sprint"
authors = ["Борис Гельруд <agu-66@mail.ru>", "Косянов Никита <>", "Холкин Сергей <>"]
license = "open"
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.10"
redis = "^5.0.1"
fastapi = "^0.104.1"
pydantic = {extras = ["email"], version = "^2.5.3"}
orjson = "^3.9.10"
pydantic-settings = "^2.1.0"
gunicorn = "^21.2.0"
uvicorn = "^0.24.0.post1"
aiologger = "^0.7.0"
aiofiles = "^0.7.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
werkzeug = "^3.0.1"
pycryptodomex = "^3.20.0"
bcrypt = "^4.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlalchemy-utils = "^0.41.1"
python-multipart = "^0.0.6"
itsdangerous = "^2.1.2"
aiokafka = "^0.10.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
